+++ lib/index.js
@@ -1,6 +1,7 @@
 const gatherTelemetryForUrl = require('./gather/gather-telemetry');
 const { getModulePathFor } = require('./utils/get-module-path-for');
 const { getTelemetry, setTelemetry, getTelemetryFor } = require('./utils/telemetry');
+const parseMeta = require('../lib/gather/local-gather');
 
 module.exports = {
   getTelemetry,
@@ -8,4 +9,5 @@ module.exports = {
   gatherTelemetryForUrl,
   getTelemetryFor,
   getModulePathFor,
+  parseMeta,
 };
